requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-core==1.1.1
azure-cosmos==4.0.0b6
absl-py==0.8.0
appdirs==1.4.4
astor==0.8.0
astroid==2.0.4
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.0
decorator==4.4.0
distlib==0.3.0
docutils==0.14
filelock==3.0.12
gast==0.3.2
google-pasta==0.1.7
grpcio==1.23.0
gevent
h5py==2.10.0
idna==2.8
imageio==2.5.0
importlib-metadata==1.6.0
isort==4.3.4
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.3.1
Markdown==3.1.1
mccabe==0.6.1
networkx==2.3
nltk==3.4
numpy==1.15.3
opencv-python==3.4.9.33
Pillow==6.0.0
protobuf==3.9.2
py==1.7.0
pycodestyle==2.4.0
Pygments==2.3.0
pylint==2.1.1
PyWavelets==1.0.3
PyYAML==5.1.2
requests==2.24
singledispatch==3.4.0.3
six==1.11.0
termcolor==1.1.0
tensorflow==1.15.3
urllib3==1.24.1
virtualenv==20.0.21
wrapt==1.11.2
zipp==3.1.0
mailjet_rest==1.3.3
azure-devops==5.0.0b5